+++ src/docx/file.ts
@@ -1,10 +1,11 @@
 import { Media } from "../media";
-import { Numbering } from "../numbering";
 import { IPropertiesOptions, Properties } from "../properties";
 import { Styles } from "../styles";
 import { DefaultStylesFactory } from "../styles/factory";
 import { Document } from "./document";
-import { Paragraph, Table } from "./index";
+import { Numbering } from "./numbering";
+import { Paragraph } from "./paragraph";
+import { Table } from "./table";
 
 export class File {
