+++ src/docx/index.ts
@@ -2,3 +2,4 @@ export * from "./document";
 export * from "./paragraph";
 export * from "./table";
 export * from "./file";
+export * from "./numbering";
