+++ src/docx/numbering/level.ts
@@ -1,8 +1,8 @@
-import * as paragraph from "../docx/paragraph/formatting";
-import { ParagraphProperties } from "../docx/paragraph/properties";
-import * as formatting from "../docx/paragraph/run/formatting";
-import { RunProperties } from "../docx/paragraph/run/properties";
-import { Attributes, XmlAttributeComponent, XmlComponent } from "../docx/xml-components";
+import * as paragraph from "../paragraph/formatting";
+import { ParagraphProperties } from "../paragraph/properties";
+import * as formatting from "../paragraph/run/formatting";
+import { RunProperties } from "../paragraph/run/properties";
+import { Attributes, XmlAttributeComponent, XmlComponent } from "../xml-components";
 
 interface ILevelAttributesProperties {
     ilvl?: number;
