+++ src/docx/paragraph/paragraph.spec.ts
@@ -2,7 +2,7 @@ import { assert, expect } from "chai";
 
 import * as docx from "../../docx";
 import { Formatter } from "../../export/formatter";
-import { Numbering } from "../../numbering";
+import { Numbering } from "../numbering";
 
 describe("Paragraph", () => {
     let paragraph: docx.Paragraph;
