+++ src/docx/paragraph/paragraph.ts
@@ -1,6 +1,6 @@
 // http://officeopenxml.com/WPparagraph.php
 import { IData } from "../../media/data";
-import { Num } from "../../numbering/num";
+import { Num } from "../numbering/num";
 import { XmlComponent } from "../xml-components";
 import { PictureRun, Run, TextRun } from "./run";
