+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from "./docx";
 export * from "./export";
-export { Numbering } from "./numbering";
 export { Styles } from "./styles";
 export { Media } from "./media";
 export * from "./export";
